Removed redundant in_check inside unmake_move for a ~1.4x speedup

+++ src/timer.hpp
@@ -1,7 +1,7 @@
 #ifndef TIMER_HPP
 #define TIMER_HPP
 
-extern int elapsed;
+extern double elapsed;
 
 void start_time();
 void end_time();
